refactor: add Abstract class to test Actor using BaseDB instances
