+++ tests/unit/puller/test_puller_actor.py
@@ -39,14 +39,7 @@ from powerapi.report import Report
 from powerapi.puller import PullerActor
 
 from ..actor.abstract_test_actor import AbstractTestActor
-from ..db_utils import FakeDB
-
-
-def define_database_content(content):
-    def wrap(func):
-        setattr(func, '_content', content)
-        return func
-    return wrap
+from ..db_utils import FakeDB, AbstractTestActorWithDB, define_database_content, REPORT1, REPORT2
 
 def pytest_generate_tests(metafunc):
     """
@@ -72,11 +65,7 @@ class FakeDispatcher:
     def send_data(self, report):
         self.q.put(report, block=False)
 
-class TestPuller(AbstractTestActor):
-    
-    @pytest.fixture
-    def fake_db(self, content):
-        return FakeDB(content)
+class TestPuller(AbstractTestActorWithDB):
 
     @pytest.fixture
     def fake_dispatcher(self):
@@ -90,10 +79,8 @@ class TestPuller(AbstractTestActor):
         fake_filter.get_type = Mock(return_value=Report)
         return PullerActor('puller_test', fake_db, fake_filter, 0, level_logger=logging.DEBUG)
 
-    def test_starting_actor_make_it_connect_to_database(self, started_actor, fake_db):
-        assert fake_db.q.get(timeout=2) == 'connected'
 
-    @define_database_content([Report(1, 2, 3), Report(4, 5, 6)])
+    @define_database_content([REPORT1, REPORT2])
     def test_start_actor_with_db_thath_contains_2_report_make_actor_send_reports_to_dispatcher(self, started_actor, fake_dispatcher, content):
         for report in content:
             assert fake_dispatcher.q.get(timeout=2) == report
@@ -101,4 +88,3 @@ class TestPuller(AbstractTestActor):
     def test_starting_actor_in_stream_mode_make_it_termiante_itself_after_empty_db(self, started_actor):
         started_actor.join(2)
         assert started_actor.is_alive() is False
-
